Edited to fix Issue #7
A problem that I was experiencing is that when the update for the FC menu was happening the menu would pop back up to the top of the screen i.e. it would scroll back up so I could not see all options.

This fix eliminates that issue and I also believe that we may see some speed improvements but that is yet to be proven.