Updates for CC 2.002
Golden cookies have been made a little more generic in CC. Golden
cookies and reindeer are now a type of "shimmer". This was causing
some errors on load, but the worse ones were in the reverse function
which was not properly quoting the string for the current season.
Both of these errors caused hangs on load, but there are more things
that need to be corrected, including the reindeer code.